Fix win32 example server writes to const nodes
The zcoap-server has recently changed its URI tree from using const URI
nodes to read/write nodes.  The purpose of this is to be able to store
subscription data in the nodes for observables.

However, the win32 example server was not updated accordingly and still
has const nodes in its URI tree.  Thus, on certain platforms,
dereference of the const nodes as non-const causes a crash.  This
commit fixes the issue.